feat: speed up `LookUp()`
Previously, `CommonEntry` is grouped by the lower bits of the board
key, and entries are sorted by the higher bits of the key. The data
structure made searching speed slower because entries in the cluster
should always be sorted.

Now, `TTCluster` is removed, and `CommonEntry` is stored in a simple
vector-like data structure.